Refactor complex midstream code
