--- pkg/base/testdata/base-specs/pod.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: base-test-pod
-  namespace: base-test
-spec:
-  initContainers:
-  - image: quay.io/replicatedcom/qa-kots-3:alpine-3.6
-    name: private-app-image
-  - image: quay.io/replicatedcom/someimage:1@sha256:25dedae0aceb6b4fe5837a0acbacc6580453717f126a095aa05a3c6fcea14dd4
-    name: private-app-image-with-digest
-  - image: redis:7@sha256:e96c03a6dda7d0f28e2de632048a3d34bb1636d0858b65ef9a554441c70f6633
-    name: public-app-image-with-digest
-  containers:
-  - image: busybox
-    name: busybox-container
