--- pkg/base/testdata/replicated-registry/pod.yaml
@@ -1,14 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: replcated-registry-test-pod
-  namespace: replicated-registry-test
-spec:
-  containers:
-  - image: registry.replicated.com/appslug/image:version
-    name: replicated-registry-image
-  - image: my-registry.example.com/appslug/some-other-image:version
-    name: custom-registry-hostname-image
-  - image: quay.io/replicatedcom/someimage:1
-    name: private-image
-
